detector/requirements.txt
opencv-python==4.5.1.48
optuna==2.8.0
pandas==1.1.5
Pillow==8.2.0
pytorch-lightning==1.2.10
pytorch-model-summary==0.1.2
pytorchtools==0.0.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
torch==1.8.1
torchmetrics==0.2.0
torchsummary==1.5.1
torchvision==0.9.1
tqdm==4.60.0
transformers==4.5.1